update to use BXML instead of REST
