+++ routes/bandwidth_routes.js
@@ -1,19 +1,19 @@
 const express = require('express');
 let router = module.exports = express.Router();
-const bw = require('../controllers/bw_controller.js');
+const bxml = require('../controllers/bxml_controller.js');
 const db = require('../controllers/db_controller.js');
 const debug = require('debug')('sip-handler');
 
 router.route('/messages')
-	.post(
-		bw.validateMessage,
-		bw.sendMessage
+	.get(
+		bxml.validateMessage,
+		bxml.sendMessage
 		);
 
 router.route('/calls')
-	.post(
-		bw.validateCall,
-		bw.checkEvent,
+	.get(
+		bxml.validateCall,
+		bxml.checkEvent,
 		db.findNumbers,
-		bw.transferCall
+		bxml.transferCall
 		);
